# Gain up to 2% speed on Intel Silvermont & Haswell processors.